Added logout and cleaned up code
