+++ frontend/src/pages/Login.jsx
@@ -6,65 +6,59 @@ import { useEffect } from "react";
 import { useNavigate } from "react-router-dom";
 
 export default function Login() {
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [loading, setLoading] = useState(false);
-  const navigate = useNavigate();
-  //Auth Context
-  const { session, login } = useAuth();
+	const [email, setEmail] = useState("");
+	const [password, setPassword] = useState("");
+	const [loading, setLoading] = useState(false);
+	const navigate = useNavigate();
+	const { session, login } = useAuth();
 
-  //redirect to homepage when already authenticated
-  useEffect(() => {
-    if (session) {
-      //set to homepage
-      navigate("/");
-    }
-  }, [session]);
+	useEffect(() => {
+		if (session) {
+			navigate("/dashboard");
+		}
+	}, [session]);
 
-  async function handleSubmit(e) {
-    e.preventDefault();
+	async function handleSubmit(e) {
+		e.preventDefault();
 
-    try {
-      //disable submit button
-      setLoading(true);
-      await login(email, password);
-      //TODO dashboard page protected routes
-      navigate("/dashboard");
-    } catch (e) {
-      //TODO catch error
-      alert("Failed to login");
-    }
-    setLoading(false);
-  }
+		try {
+			setLoading(true);
+			await login(email, password);
+			navigate("/dashboard");
+		} catch (e) {
+			alert("Failed to login");
+		}
+		setLoading(false);
+	}
 
-  return (
-    <div>
-      <h2>Login to Your Account</h2>
-      <form onSubmit={handleSubmit}>
-        <label htmlFor="email">Email Address:</label>
-        <input
-          id="email"
-          name="email"
-          type="email"
-          autoComplete="email"
-          placeholder="Email Address"
-          onChange={(e) => setEmail(e.target.value)}
-          required
-        />
-        <label htmlFor="password">Password:</label>
-        <input
-          id="password"
-          name="password"
-          type="password"
-          placeholder="Password"
-          onChange={(e) => setPassword(e.target.value)}
-          required
-        />
-        <button type="submit" disabled={loading}>
-          Login
-        </button>
-      </form>
-      <Link to="/register">Don't have an Account? Register</Link>
-    </div>
-  );
+	return (
+		<div>
+			<h2>Login to Your Account</h2>
+			<form onSubmit={handleSubmit}>
+				<label htmlFor="email">Email Address:</label>
+				<input
+					id="email"
+					name="email"
+					type="email"
+					autoComplete="email"
+					placeholder="Email Address"
+					onChange={(e) => setEmail(e.target.value)}
+					required
+				/>
+				<label htmlFor="password">Password:</label>
+				<input
+					id="password"
+					name="password"
+					type="password"
+					placeholder="Password"
+					onChange={(e) => setPassword(e.target.value)}
+					required
+				/>
+				<button type="submit" disabled={loading}>
+					Login
+				</button>
+			</form>
+			<Link to="/register">Don't have an Account? Register</Link>
+		</div>
+	);
 }
